Add user integration test and fix buggs

+++ apps/backend/src/main/java/ch/heig/pdg/backend/services/UserService.java
@@ -3,6 +3,7 @@ package ch.heig.pdg.backend.services;
 import ch.heig.pdg.backend.dto.UserDTO;
 import ch.heig.pdg.backend.dto.mapping.UserMapper;
 import ch.heig.pdg.backend.entities.User;
+import ch.heig.pdg.backend.exception.exceptions.ForbiddenOperationException;
 import ch.heig.pdg.backend.repositories.InventoryRepository;
 import ch.heig.pdg.backend.repositories.UserRepository;
 import ch.heig.pdg.backend.security.utils.CurrentUser;
@@ -10,6 +11,7 @@ import ch.heig.pdg.backend.utils.HugoSearchFilter;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 @Service
@@ -34,6 +36,9 @@ public class UserService extends AbstractService {
     }
 
     public UserDTO removeUser(Integer id) {
+        if (!Objects.equals(this.currentUser.getCurrentUserData().userId, id)) {
+            throw new ForbiddenOperationException("Can not remove other user");
+        }
         UserDTO userDTO = (UserDTO) this.userMapper.getDTO(this.getEntityIfExists(id, this.userRepository));
         this.userRepository.deleteById(id);
         return userDTO;
@@ -54,6 +59,9 @@ public class UserService extends AbstractService {
     }
 
     public UserDTO updateUser(Integer id, UserDTO userDTO) {
+        if (!Objects.equals(this.currentUser.getCurrentUserData().userId, id)) {
+            throw new ForbiddenOperationException("Can not update other user");
+        }
         return (UserDTO) this.userMapper.getDTO(
                 this.userRepository.save(
                         this.userMapper.updateFromDTO(
